refactor(cli): simplify telemetry code
