+++ apps/cli/src/config/index.ts
@@ -11,6 +11,7 @@ import { promptNumber, promptConfirm } from '../utils/user-input';
 import { log, configureLogger } from '../utils/logger';
 import { tokenManager } from '../auth/token-manager';
 import { oauthManager } from '../auth/oauth';
+import { analyticsEvents } from '../utils/telemetry';
 
 export class ConfigResolver {
   private config: Config | null = null;
@@ -74,7 +75,6 @@ export class ConfigResolver {
           log.info(`Configuration saved to ${CONFIG_FILE_NAME}`);
 
           // Track config storage event
-          const { analyticsEvents } = await import('../analytics/events');
           await analyticsEvents.storedConfigJson();
         }
       } else if (!configFile || configFile.appPort !== appPort) {
@@ -220,7 +220,6 @@ export class ConfigResolver {
         log.info(`Configuration saved to ${CONFIG_FILE_NAME}`);
 
         // Track config storage event
-        const { analyticsEvents } = await import('../analytics/events');
         await analyticsEvents.storedConfigJson();
       }
     }
