+++ apps/cli/src/index.ts
@@ -1,8 +1,6 @@
 import { configResolver } from './config';
 import { configFileExists } from './config/config-file';
-import { posthog } from './analytics/posthog';
-import { telemetryManager } from './config/telemetry';
-import { analyticsEvents } from './analytics/events';
+import { telemetryManager, analyticsEvents } from './utils/telemetry';
 import { identifierManager } from './utils/identifier';
 import { getServer } from './server';
 import { shutdownAgent } from './server/agent-loader';
@@ -145,12 +143,12 @@ async function main() {
     await identifierManager.getMachineId();
 
     // Initialize analytics after config is resolved
-    await posthog.initialize();
+    await telemetryManager.initialize();
 
     // Set user properties if authenticated
     const authState = await oauthManager.getAuthState();
     if (authState?.isAuthenticated) {
-      posthog.setUserProperties({
+      telemetryManager.setUserProperties({
         user_id: authState.userId,
         user_email: authState.userEmail,
       });
@@ -263,8 +261,8 @@ async function main() {
       try {
         await analyticsEvents.cliShutdown();
 
-        // Shutdown PostHog client
-        await posthog.shutdown();
+        // Shutdown telemetry
+        await telemetryManager.shutdown();
       } catch (_error) {
         // Ignore analytics errors during shutdown
       }
